requirements.txt
Django==4.2.9
mypy==1.8.0
flake8==7.0.0
djangorestframework==3.14.0
